# backend/src/main/java/com/examarchitect/backend/auth/dto/LoginRequest.java
package com.examarchitect.backend.auth.dto;

import jakarta.validation.constraints.Email;
import jakarta.validation.constraints.NotBlank;

public record LoginRequest(
    @Email(message = "Invalid email address")
    @NotBlank(message = "Email is required")
    String email,

    @NotBlank(message = "Password is required")
    String password
) {
}